settings page

+++ src/components/Search/Search.jsx
@@ -8,10 +8,11 @@ import { useGlobalState } from '../../Context'
 import { SearchCity } from '../../api/SearchCity'
 import { CityWeather } from '../../api/CityWeather'
 import { sortByCity } from '../../utilities/Sorter'
+import _ from 'lodash'
 
 const Search = () => {
   //when searching, make sure before you go to a city page, check if that city is in favorites, then go to the city so it can show its notes
-  const {searchValue:[searchText,setSearchText], globalLoadingValue:[globalLoading, setGlobalLoading], favoritesValue:[favoritesList, setFavoritesList], currentCityValue:[currentCity, setCurrentCity]} = useGlobalState()
+  const {searchValue:[searchText,setSearchText], globalLoadingValue:[globalLoading, setGlobalLoading], favoritesValue:[favoritesList, setFavoritesList], currentCityValue:[currentCity, setCurrentCity], exploreValue:[exploreList, setExploreList]} = useGlobalState()
   const [isSearching, setIsSearching] = useState(false)
   const [searchResults, setSearchResults] = useState([])
   const [isLoading, setIsLoading] = useState(true)
@@ -43,15 +44,45 @@ const Search = () => {
     if(isOnFavoritesPage){
       clearSearch();
       setGlobalLoading(true)
-      let res = await CityWeather(obj.coord)
-      let newObj = {...res, notes:[]}
-      setFavoritesList((prevList)=>{
-        //sort here
-        let newList = [...prevList]
-        newList.push(newObj)
-        localStorage.setItem('RESA_favorites', JSON.stringify(sortByCity(newList)))
-        return newList
-      })
+      // check if the city already exists in favorites list or the explore list
+      const existingFavIndex = favoritesList.findIndex(item => item.coord === obj.coord)
+      const existingExpIndex = exploreList.findIndex(item => item.coord === obj.coord)
+      if(existingFavIndex === -1 && existingExpIndex === -1){
+        //if city does not exist in both lists, then add the new city
+        let res = await CityWeather(obj.coord)
+        let newObj = {...res, notes:[]}
+        setFavoritesList((prevList)=>{
+          let newList = [...prevList]
+          newList.push(newObj)
+          localStorage.setItem('RESA_favorites', JSON.stringify(sortByCity(newList)))
+          return sortByCity(newList)
+        })
+        console.log('city is entirely new')
+      }
+      else if(existingFavIndex === -1 && existingExpIndex !==-1){
+        //city exists in the explore list already
+        let existingCity = _.cloneDeep(exploreList[existingExpIndex])
+        //remove it from explore
+        setExploreList((prevList)=>{
+          let newList = [...prevList]
+          newList.splice(existingExpIndex,1)
+          localStorage.setItem('RESA_explore', JSON.stringify(newList))
+          return newList
+        })
+        //update it with notes and put it in favorites
+        let updatedExistingCity = {...existingCity, notes:[]}
+        setFavoritesList((prevList)=>{
+          let newList = [...prevList]
+          newList.push(updatedExistingCity)
+          localStorage.setItem('RESA_favorites', JSON.stringify(sortByCity(newList)))
+          return sortByCity(newList)
+        })
+        console.log('city existed in explore')
+
+      }
+      else{
+        console.log('CITY already exits in favorites')
+      }
       setGlobalLoading(false)
     }
     else{
@@ -59,9 +90,17 @@ const Search = () => {
       clearSearch()
       setGlobalLoading(true)
       browse('/city')
-      let res = await CityWeather(obj.coord)
-      setCurrentCity(res)
-      localStorage.setItem('RESA_currentCity', JSON.stringify(res))
+      //check if the city already exists in favorites list (so that user will have access to their notes)
+      const existingFavIndex = favoritesList.findIndex(item => item.coord === obj.coord)
+      if(existingFavIndex !== -1){
+        setCurrentCity(favoritesList[existingFavIndex])
+        localStorage.setItem('RESA_currentCity', JSON.stringify(favoritesList[existingFavIndex]))
+      }
+      else{
+        let res = await CityWeather(obj.coord)
+        setCurrentCity(res)
+        localStorage.setItem('RESA_currentCity', JSON.stringify(res))
+      }
       setGlobalLoading(false)
     }
   }
